+++ .rubocop.yml
@@ -35,13 +35,12 @@ Metrics/ClassLength:
     - 'app/services/cocina/from_fedora/descriptive/admin_metadata.rb'
     - 'app/services/cocina/to_fedora/descriptive/title.rb'
     - 'app/services/cocina/from_fedora/descriptive/titles.rb'
-    - 'app/services/cocina/mods_normalizer.rb'
+    - 'app/services/cocina/normalizers/mods_normalizer.rb'
     - 'app/services/mods_equivalent_service.rb'
     - 'app/services/cocina/from_fedora/descriptive/identifier_type.rb'
     - 'app/services/cocina/to_fedora/descriptive/contributor_writer.rb'
     - 'app/services/cocina/to_fedora/descriptive/form.rb'
-    - 'app/services/cocina/mods_normalizers/subject_normalizer.rb'
-    - 'app/services/cocina/mods_normalizers/origin_info_normalizer.rb'
+    - 'app/services/cocina/normalizers/mods/subject_normalizer.rb'
     - 'app/services/cocina/from_fedora/descriptive/access.rb'
     - 'app/services/cocina/to_fedora/descriptive/related_resource.rb'
     - 'app/services/cocina/object_updater.rb'
