+++ spec/services/cocina/normalizers/mods/geo_extension_normalizer_spec.rb
@@ -2,8 +2,8 @@
 
 require 'rails_helper'
 
-RSpec.describe Cocina::ModsNormalizers::GeoExtensionNormalizer do
-  let(:normalized_ng_xml) { Cocina::ModsNormalizer.normalize(mods_ng_xml: mods_ng_xml, druid: 'druid:pf694bk4862', label: nil).to_xml }
+RSpec.describe Cocina::Normalizers::Mods::GeoExtensionNormalizer do
+  let(:normalized_ng_xml) { Cocina::Normalizers::ModsNormalizer.normalize(mods_ng_xml: mods_ng_xml, druid: 'druid:pf694bk4862', label: nil).to_xml }
 
   context 'when normalizing geo PURL' do
     let(:mods_ng_xml) do
