+++ spec/validators/cocina/description_roundtrip_validator_spec.rb
@@ -102,7 +102,7 @@ RSpec.describe Cocina::DescriptionRoundtripValidator do
 
     context 'when invalid' do
       before do
-        allow(Cocina::ModsNormalizer).to receive(:normalize).and_return(Nokogiri::XML(mods.gsub('Chi Running', 'Zen of Running')))
+        allow(Cocina::Normalizers::ModsNormalizer).to receive(:normalize).and_return(Nokogiri::XML(mods.gsub('Chi Running', 'Zen of Running')))
       end
 
       it 'returns failure' do
